Rename device module to system. Rename KiwiTalkSystemInfo to SystemInfo. Move SystemInfo initialization to system module.
